Add thk threshold in interp (needed, see below).
This is needed because if masking is done beforehand (using xarray.where as we
do now), values near the margin will be replaced by nan, and interpolation near
the margin will become impossible. If masking is done after interp on the other
hand, interp will not be able to interpolate the mask between 0 and 1, and to
compute a detailed masked based on the new topography.

This issue could actually be worked around by using a separate variable for the
ice mask (standard name land_ice_area_fraction). This variable will be
automatically interpolated, and other variables will not be filled with nans.
In any case, masking seem necessary alongside interp, thus I find it convenient
to have the threshold parameter at hand in the interp function.